Fix incorrect name lookup for decorated methods (#8175)
Resolves #8161 

According to comments of `lookup` in `mypy/semanal.py`, when we look up a class attribute, we require that it is defined textually before the reference statement, thus line number is used for comparison.

When function has decorators, its line number is determined by the top decorator instead of the `def`. That's why #8161's code fails because on line 8, the `A` in `Type[A]` has the line number of 8 while the `@staticmethod` function `A` has the line number of 7 due to the decorator. 

Thus we need to properly handle this by introducing the number of decorators when deciding textural precedence. Also overloads needs special handling to be considered "as a unit".