dataclasses: Fix deeply nested InitVar definitions with init=False (#8208)
Fixes #8207

In #8159, I fixed InitVar handling so that it looked up the superclass __init__ 
and used that to determine the definition of InitVars. That fix doesn't work 
in the presence of `init=False` subclasses of the initial `init=False` class
In that case, the lookup fails because the first attribute reached in MRO 
order on the subclass is the intermediate class, which also doesn't have 
the variable in its `__init__`. Then mypy didn't know about the InitVar but 
still tried to process it as if it did, resulting in an assertion error.

This PR fixes that issue by not adding the field to the set of known attrs 
until an actual definition is found.