Allow redefining TypedDict keys while still throwing an error (#8109)
Creating a TypedDict type that is similar to an existing one but has different 
types for one or two keys was impossible and required redefining the 
TypedDict completely, which can be very verbose in the case of complex types. 
Throwing a type error in the same way as before, but allowing the overwrite to 
go through, seems much more reasonable for quality of life purposes.