Update docs about variables and aliases (#8200)
Resolves #3494 (Since module types are tracked in #3500)

Following #8187 (comment) and #3494, if I understand correctly and the semantics 
in #3494's example code has been fixed (report error on re-assign Alias = B, the 
remaining work of #3494 is to update the docs, which is the main focus of this PR.

Newly added docs are in common issues and solutions section, with the content mostly adapted from Ivan's example in #3494. And a note point to the docs is also added.