[mypyc] Refactor: extract non-local control classes from mypyc.genops (#8406)
This is the first PR in my quest to tidy up mypyc.genops.

My goal is to eventually break it up into files no longer than 1500 lines
each, and with a clean dependency structure. In the initial stages there
will be many cyclic dependencies, but I plan to fix those later on.