[mypyc] Refactor: remove cyclic deps by extracting genops visitor (#8430)
Add a simple visitor module that doesn't contain any implementation and just
dispatches to various transform implementations. The main benefit is that
the implementations don't need to depend on the visitor (and thus also don't
need to depend on the other implementations) and that `IRBuilder` doesn't
need to depend on the implementations.

Also extract the driver function to another module, as this would otherwise
cause an import cycle.

Some cyclic dependencies remain and wouldn't be hard to get rid of. I
wanted to keep this PR relatively simple so I didn't fix them yet.

Work on mypyc/mypyc#714.