Fix bug by removing fix_suppressed_dependencies
We found a bug internally caused by deleting an indirect dependency
with --follow-imports=skip and --warn-unused-ignores. This is caused
by the indirect dependency being moved from deps to suppressed by
fix_suppressed_dependencies, which loses the knowledge that it is an
indirect dependency, causing verify_dependencies to take it much more
seriously than it should be.

fix_suppressed_dependencies was added as a hotfix for an incremental
bug discovered internally in 2016 that we never found a test case
for. I hypothesized during an earlier refactor that it had been made
redundant. Since fix_suppressed_dependencies is causing trouble now,
let's put that hypothesis to the test and remove it. If this causes a
regression, then maybe we'll be able to finally extract the test case
we want.