Docs: use asterisk notation when ignoring library
The examples for ignoring missing imports for a specific
library are currently conflicting between:

https://mypy.readthedocs.io/en/latest/running_mypy.html#missing-type-hints-for-third-party-library
and
https://mypy.readthedocs.io/en/stable/existing_code.html#start-small

The former uses the syntax ``[mypy-library]``.

while the latter uses the syntax ``[mypy-library.*]``.

It's the latter that captures the goal of what 99% of users are trying
to do, which is to ignore a package.

This is made clearer with a concrete example using the Ansible
package, broken into stages.

In stage 1 we have:

    $ cat foo.py
    import ansible

    import ansible.constants

    $ cat mypy.ini
    [mypy]
    warn_unreachable = True

    [mypy-ansible]
    ignore_missing_imports = True

This mirrors the current ``[mypy-foobar]``.

In this case mypy will still raise for ``ansible.constants``:

    $ mypy --python-version '2.7' --platform linux --show-traceback .
    foo.py:3: error: Skipping analyzing 'ansible.constants': found module but no type hints or library stubs
    foo.py:3: note: See https://mypy.readthedocs.io/en/latest/running_mypy.html#missing-imports

This is almost certainly _not_ what most users want.  If they don't
have stubs for the Ansible package, they don't have them--period.
The language specifically uses "library," which is best understood
to be a package rather than a single module.

Stage 2:

    $ cat mypy.ini
    [mypy]
    warn_unreachable = True

    [mypy-ansible.*]
    ignore_missing_imports = True

    $ mypy --python-version '2.7' --platform linux --show-traceback .
    Success: no issues found in 1 source file

The asterisk version ignores Ansible and its modules/subpackages.