Fix potential wrong-file error message for unsupported Types
If we want to report error messages at the use of a type, we have to
report it when processing that type, not when we run into some code
pattern that uses it, or else we can generate unignorable misplaced
error messages.

It doesn't seem worth bothering to do the refactors that this would
require so I just moved the error messages to the use-sites and
adjusted the text to be more appropriate.

Fixes #8825.