[mypyc] Add a primitive for formatting an int as a str
This is partially just shameless cheesing on an example kmod used in a
blog post (http://blog.kevmod.com/2020/05/python-performance-its-not-just-the-interpreter/comment-page-1/#comment-55896),
but it is generally useful. Probably it will be more useful as part of
more general string formatting optimizations, but it is a start.

This gives a 3x speedup on the example in the blog post.

I had been hoping to just allocate a unicode string and call sprintf
into it but that turned out to be quite a bit slower than boxing the
int and calling PyObject_Str on it! So I just implemented it.