Fix untyped decorator overload error on class decorator with __call__ overloads (#9232)
When --disallow-untyped-decorators is used, a callable-class decorator 
with a overloads on its __call__ implementation would incorrectly issue an 
"Untyped decorator makes function untyped" error, even when the overloads 
are typed properly. This has been reported in pytest, which does this: 
pytest-dev/pytest#7589.

This fixes the problem by expanding on a previous fix in this area, #5509.