Speed up make_simplified_union for Literal[string]s (issue #9169) (#9192)
make_simplified_union no longer runs for 0.5 seconds every time it's called with a 
Union containing 270 literals of strings.

Like I explained in #9169, this only fixes half of the problem and I'm not capable of 
fixing the other half. This function is still called 1098 times for the "reduced" example 
code, and IMO it should be called only once.