[mypyc] Implement LoadAddress (#9287)
This PR introduces LoadAddress, an op for taking the address of a given name.

Currently, in mypyc we only need to take the address of a name when we are 
using the name_ref_ops, so the op takes a string (CPython name) as its argument. 
We can revisit the design later if we want to take the address of arbitrary name 
later.