add narrowing for lists via index isinstance check
Allows for narrowing a union of list types by indexing into the list
and checking the type of an element.

mypy already narrows the specific index expression.
Since mypy already narrows the index expression's type, we use the type
to compare against the type arguments of the lists and generate a new
type that is narrower than the original.

```
from typing import Union, List

a: Union[List[List[int]], List[int]]

# before
if isinstance(a[0], list):
    reveal_type(a) # N: Revealed type is 'Union[builtins.list[builtins.list[builtins.int]], builtins.list[builtins.int]]'
else:
    reveal_type(a) # N: Revealed type is 'Union[builtins.list[builtins.list[builtins.int]], builtins.list[builtins.int]]'

# after
if isinstance(a[0], list):
    reveal_type(a) # N: Revealed type is 'builtins.list[builtins.list[builtins.int]]'
else:
    reveal_type(a) # N: Revealed type is 'builtins.list[builtins.int]'
```

fixes: https://github.com/python/mypy/issues/9362