Fix search paths for `-p`
`-p` is similar to `-m` except that it will recursively typecheck submodules.

Unfortunately the early module walk was being done with search paths that
did not benefit from site_packages for the selected python executable,

which resulted in some packages being typechecked fine with `-m`, but
rejected as `not found` by `-p`.

Fixes #9386