[mypyc] Try to fix mypy mac build failure on Python 3.5 on Travis
We may be building a mixed 32/64-bit binary, so restrict the maximum integer
literal to what's supported on a 32-bit platform, since the same C code
needs to compile on both 32-bit and 64-bit architectures.

This will make some slicing operations a bit slower in 32/64-bit mode
(Python 3.5 on mac).