Require first argument of namedtuple to match with variable name (#9577)
Closes #4589

This PR modifies check_namedtuple to return the internal name of the namedtuples (e.g. the content of the first argument of namedtuple/NamedTuple) so that the callers, especially analyze_namedtuple_assign, can check if the name of the variable on the l.h.s. matches with the first argument of the namedtuple.