teststubtest: drop StubtestIntegration
I did some work to speed up stubtest tests substantially in #9621
This is a lesser win. The argument to get rid of it is that it's the
slowest mypy test and just testing for crashes on typeshed doesn't
provide much value. Since this was written we started running stubtest
in typeshed CI and in practice all changes are tested on typeshed
anyway.