fastparse: fix type checking on 3.8
This got broken by https://github.com/python/typeshed/pull/4740
It took me a little bit to figure out how this evaded all our CI, but
it's because we type check with version 3.5 and mypyc only with 3.5 and
3.7. That is, we do not type check with 3.8 or later and so do not check
against stdlib's AST types.

mypyc wheels did break through, but I'm not sure anyone gets to see
those. Maybe we should add a badge for them to the README.

I only noticed this because I was trying to get mypy_primer to mypyc
compile sometimes.