allow packages/modules as args with files in cfg
Currently only files can be specified on the command line if files are
specified in mypy.ini.  This looks a great deal like a bug, especially
since packages and modules cannot be specified in the configuration
file.  This commit changes the logic to only use the files from the ini
file if none of (modules/packages/files) are specified on the command
line and adds tests to verify the new behavior.