[mypyc] Remove Environment (#9840)
Remove `mypyc.ir.ops.Environment`. Represent potentially uninitialized variables 
explicitly in the IR by generating an assignment of an error value, instead of relying 
on the environment to keep track of uninitialized variables.

This simplifies the IR and makes it more consistent, and it will also make it a little
easier to create new backends. More importantly, this removes one variant of
the "environment" concept. The term is less overloaded now.

Work on mypyc/mypyc#781.