Add a class attribute hook to the plugin system
This adds a hook to modify attributes on *classes* (as opposed to the existing attribute hook, which is for attributes on *instances*). This also adds a test to demonstrate the new behavior. The modifications and added tests were modeled off of https://github.com/python/mypy/commit/3acbf3fe78a61c19ff96754233ada453472004c4. This is intended to solve https://github.com/python/mypy/issues/9645