Return mypy_primer to CI (#9842)
mypy_primer was temporarily removed in #9769. This brings it back.
See https://github.com/python/typeshed/pull/4806 for the equivalent
typeshed PR.

One change is that we now shard across three jobs for speed. We also no
longer mypyc compile, since we should be right about the point where the
gains from compilation are outweighted by compilation time (for
`--mypyc-compile-level 0`). Although if there were an easy way to share
the compiled mypy across mypy runs it would be worth it...

Co-authored-by: hauntsaninja <>