add modules and packages keys to config
This adds configuration file support for specifying packages and
modules.  I tried to preserve existing semantics between the config file
and command line.  If the command line specifies anything, it overrides
the config file, if the config file contains both package/module and
files keys, the later key is ignored.

Also add tests and basic docs for the new keys.