Keep track of fine-grained deps for modules in typeshed
Even if we assume that typeshed stubs don't change, we need the deps
when bringing in new stubs when following imports during a fine-grained
increment.

This fixes false positives when following imports to `six` and
`requests`.

Some modules are always part of the initial build, and we can still
skip generating deps for them, as an optimization.

I only tested this manually, since I couldn't figure out an easy way to
reproduce the issue in our tests.

Fixes #10033.