update --warn-unreachable to respect if the code should be type checked
Mypy currently will warn about unreachable code regardless if the code
"should" be type checked or not. Adding `reveal_type` to a function would
then always produce `Any` which is unhelpful and confusing. This change
updates mypy to only warn about unreachable code if the code is in a
checked function.