Add type is not equal tests
Add tests to make sure that type(x) != T and type(x) is not T work as
expected (the same as type(x) == T and type(x) is T but with the if and
else branches switched.

Currently the type(x) != T check is failing because of a mypy error
about an unsupported left operand type.