Narrow types in else case if type is final
Final types cannot be subclassed, so it's impossible for a subclass of a
final type to be used in the else case of a comparison of type(x) to a
final type. That means we can narrow types in the else case the same way
we would do for isinstance checks if type(x) is being compared to a
final type.