Fix KeyError when looking up type info
lookup_typeinfo appears to crash with a KeyError if it can't find the
class corresponding to the given name, which becomes a problem when
we're trying to check if the type someone is comparing to is final.

This fixes that by assuming any types or classes we can't look up are
not final (so that we don't narrow types in the else case when it
doesn't make sense to).