Improve check if class is final
Using lookup_typeinfo to get the TypeInfo of a class by name results in
a lot of different exceptions being thrown if what we're passing in
isn't actually the name of a class.

This replaces the use of lookup_typeinfo by using `expr.node` to get
more info about the class that the given name is referring to.