Fix inference of IntEnum value attribute type (Python 2)
The enum stubs for Python 2 are slightly different from Python 3.
The fix in #10412 didn't quite work in Python 2 mode, since `IntEnum`
doesn't define `__new__`, and the `__new__` from `builtins.int` is
found instead. Work around this by skipping any `__new__` methods
found in built-in types.

I gave up trying to write a test case for this. It's tricky because
`enum` is in Python 3 stdlib but not in Python 2 stdlib.