Update EADME.md, MIT License link is not working
When I was going through the README.md of mypy, I saw that a link in the README.md file is not working. This is a link about the License of the project. Although all the links are working perfectly and anyone can just click on the link and see relevant informations.

The "LICENSE" should be embedded with the link  https://github.com/python/mypy/blob/master/LICENSE