Add test to reproduce CI failure
Add a test that reproduces the CI failure that occurs when trying to run
mypy after being compiled with mypyc.

This failure happens when collecting tests because mypy fails with an
AttributeError. For example, if the line in question is `from mypy
import build`, it would fail with an AttributeError saying "module
'mypy' has no attribute 'build'". This is caused by the compiled code
skipping the import call when the package being imported from (in this
case, mypy) has already been imported, leading to the code trying to
access an attribute (in this case, build) that hasn't been imported.