Remove check if package is imported in from imports
Remove the check whether the package being imported (the part before
`import`) is already loaded, instead of avoiding the import entirely if
it's already loaded as we do for regular imports.

This avoids a crash when dealing with multiple modules imported from the
same package because only the modules listed in the first import will
actually be loaded.

For example, if we have something like:
```
from pkg import mod1
from pkg import mod2
```
The second import statement will skip the import because it thinks pkg
is already imported, so attempts to access pkg.mod2 will fail with an
AttributeError.

It might be worth combining multiple from imports like the ones above
into one from import statement, but that would probably involve
analyzing a lot of separate code, which might slow down compile times,
and would probably be more trouble than it's worth.