Make ignore_missing_imports work for libs which had bundled stubs (#10582)
It has to be specified as a per-module option, if the library used
to have bundled stubs. This way mypy won't silently start ignoring
these missing stubs.

This makes ignore_missing_imports special by making it work a bit
differently, depending on whether it's set globally or per module.

The new behavior is less surprising. This makes it possible to ignore
arbitrary missing third-party stub packages, even those which used to
have bundled stubs.

Fixes #10283.

Originally implemented by @TheCleric in #10283. This is somewhat
simpler alternative implementation which is easier to reason about and
may perform better in some cases of very large configs.