Add some mypyc run tests
Add some run tests to run-singledispatch.test for situations that are
likely to be common or are likely to be edge cases that we don't handle
correctly.

Most of these tests are commented out because mypyc doesn't properly
support singledispatch yet, with the few that aren't commented out
working only because they're written in a way that causes mypyc to use
the standard library implementation of singledispatch.