[mypyc] Fix evaluation of iterable in list comprehension twice
This could result in a crash if the second evaluation results
in a shorter list, such as in this example (besides being incorrect
overall):

```
a = [s for s in f.readlines()]
```

`f.readlines()` was called twice, resulting in an empty list on
the second call. This caused the list object to have a NULL item,
which is invalid.

This should fix `mypy --install-types` in compiled mode.