Skip failing tests and fix error in test
Skip the 2 tests added in the previous commit because they're currently
failing due to lack of singledispatch support. Also removes a yield from
statement that mypy correctly points out has an incorrect type (it tries
to return an iterator of strings, but it should return an iterator of
errors according to the type signature).