Add support for using xfail in test cases (#10635)
Add support for adding -xfail to the end of a test case in a .test file
to make pytest mark it as expected to fail (so that the test is run but
the test suite doesn't fail if the test fails).

Turns on xfail_strict in the pytest config so that tests marked as xfails
that suddenly start passing get marked as failures. This makes sure that
some tests don't end up getting marked as xfails for a long time after
they started working correctly, which makes sure that they can get
turned into regular tests as soon as possible. That makes sure that we
can use those tests to catch regressions, instead of silently ignoring
when they fail.

Closes #10604