Fix faulty casts in checker (#10560)
Fixes #9615, fixes #9682

Both issue reports hit the faulty cast on L2803. This changes the logic
so that that cast is actually always true. If not, we just end up doing
whatever the fallback else clause does; this resulted in behaviour that
matched what I expected.

After that, the #9682 hit another crash in checker.py, where var.node
was None instead of a Var. It seemed reasonable to just branch instead.

Co-authored-by: hauntsaninja <>