Add failing test for assigning method to variable
Add a test that checks if plugins are correctly called with
'{class_name}.{method_name}' when a method is assigned to a local
variable and then used as a decorator, instead of the current behavior,
which is calling the plugin with '{module_name}.{local_variable_name}'
where local_variable_name is the variable that the method was assigned
to.

This is currently marked as xfail because this doesn't work right now.