Call get_method_hook when methods are used in decorators (#10675)
When a method is used in a decorator, pass the full name to plugins
through the get_method_hook function, just like we would normally do for
other places that methods got called. Previously, in these situations,
we would call get_function_hook instead with '{method_name} of
{class_name}', which loses information about which module the class is
located in, and gives information about a method to the hook meant for
functions.

Currently this is limited to situations where the method is accessed in
the decorator itself, so

    @obj.meth
    def f() -> None: ...

would work but

    method = obj.meth
    @method
    def f() -> None: ...

probably wouldn't work.