Fix several tests and remove xfails
Fix some bugs in the tests, like some of the error messages having the
wrong function name, or using `Union[..., None]` instead of
`Optional[...]` in those error messages.

That allows some of the tests that were originally marked as xfails to
start passing, so this removes the -xfail for those tests and makes them
regular tests.