Refactor add_method_to_class to allow use in type checking
Refactor add_method_to_class so that we can use most of it even when we
aren't in a semanal hook (meaning that we don't have access to a
SemanticAnalyzerPluginInterface api).

The only use of the semanal api is creating the builtins.function type,
so this creates a new function that takes that function type as an
argument so we can provide it by other means.