Modify add_method_to_class to take a CheckerPluginInterface
Change add_method_to_class to take either a CheckerPluginInterface or
SemanticAnalyzerPluginInterface instead of adding a new version of
add_method_to_class that takes a builtins.function argument. That avoids
us needing 2 versions of add_method_to_class and makes it easier to use
it from type checking plugin hooks (because we don't need to create a
builtins.function type first to use it).