Change return type of singledispatch function call to FunctionLike
Change the return type of the plugin hook for determining the signature
of a singledispatch function when it's called to FunctionLike. That
allows us to return Overloaded without needing a 'type: ignore' comment,
which should prevent mypyc from throwing an exception when it sees the
wrong type. This change should be possible now that #10717 has been
merged.