Require dispatch types to be subtype of singledispatch first argument
Currently, we type check calls to singledispatch functions by finding
all implementations that might get used and trying to make sure that the
arguments passed work for all of these implementations.

The problem with this is that we need to find all registered
implementations before we try type checking any singledispatch calls,
which is hard if you have a call to the singledispatch function inside
one of the registered implementations (This came up in stubtest.py).

This commit instead requires that the dispatch type for each registered
function is a subtype of the first argument of the main singledispatch
function, which allows us to just type check all calls as if they were
just calling the main singledispatch function, while being confident
that calls that use registered implementations work correctly.