Fix test failures
Most of the tests broke after changing the plugin to only care about the
fallback implementation when type checking calls, so this fixes those
tests.

Some of the tests are deleted entirely because they test features that
we no longer implement (like type checking non-dispatch types based on
which implementation would be used in a specific case), while others are
modified slightly to make all dispatch types subtypes of the first
argument of the singledispatch function.