Avoid storing info in metadata
The metadata that we were storing with the fallback and registered types
was stored in the TypeInfo for _SingleDispatchCallable, meaning it was
shared between all singledispatch functions, leading to some false
positives.

This fixes that by storing the fallback in the type variables, and not
storing the registered implementations at all (because it's somewhat
difficult storing a variable number of type variables).

Not storing the registered implementations unfortunately makes catching
some errors impossible, like checking whether multiple registered
implementations have the same dispatch type or if one implementation
will never be called because it's overshadowed by another one with a
more generic dispatch type.