Return function type directly from callbacks
Instead of deciding what to return in register_function, returning
ctx.default_return_type when there's an error with the register call,
just return the function directly from the callback. That allows calls
to registered functions to be type checked properly even when registered
incorrectly, and makes the code simpler and easier to test.